requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-auth==2.25.2
google-api-python-client==2.111.0
google-cloud-aiplatform==1.38.1
vertexai==0.0.1
